src/services/search_service/requirements.txt
fastapi
pydantic
httpx
uvicorn[standard]
motor
python-dotenv
boto3
python-multipart
tenacity
openai
requests
email-validator
aiohttp
PyMuPDF
pinecone

httpx